ask_embeddings.ask now accepts an optional context_query.
In many cases the query is sufficient to select related context. But in some cases the query and context are
somewhat different. E.g. "What are some rules of thumb for running a secret garden? Summarize the context
into 8 sentences or less." In that sentence, the first sentence is what is useful for selecting the context
to include, and the latter is the question but shouldn't be used to select context.

If context_query is not provided, it defaults to just being the raw question.

Part of #7.