When converting an OG library to a new one, make sure not to overshadow any entries.
The OG format allowed a list of embeddings, multiple of which might share the same issue_id. Before this
commit, only the last embedding for a given issue_id would be saved.

Now, if we detect a collision we make a new chunk_id by appending a counter until we find an unused one.

Part of #12.