Set library_for_query defaults explicitly.
It's nice to be able to call library_for_query directly passing the values from request.form.get(), and if
the value is None we want to use a default. That means we can't use the normal method parameter defaulting,
because the parameter will be explciitly set to None.

Fixes a few bugs in server, but it still doesn't work by default.