Library.query() uses sort type of manual.
This is a bit of a cheat until #52 is fixed. It ensures that the sort order is set to insertion order (query
already inserts things in sorted order). This means that the actual order is transmitted in the library.

This means that theoretically if all host endpoints were updated, get_chunk_infos_for_library wouldn't have
to resort its keys by similarity.

Part of #50.