Fix it so delete_chunk() doesn't lead to a resort every time.
Technically if the sort is random and with a seed, we could need a resort to be in canonical sort (left a
TODO).

Before, when we were deleting all of the restricted content in big libraries for people without an access
token, we were resorting every single time, which was unreasonably slow and leading to request time out.

Part of #52.