Make it so we do a bisect insertion for set_chunk in sorted libraries.
dacc3228978aabaf6eb14e146dcc30b471fb761e fixed a bug where if you deleted chunks again and again in a sorted
Library it wouldn't spend a ton of time resorting. But still technically if you did set_chunk again and
again you'd get the same problem (it's just that Library.query() didn't happen to exercise that).

This fixes it so even if you do lots of set_chunk() insertions on a sorted library we still don't do a full
resort but take advantange of the sort_ids being already sorted and just do a bisect insertion.

Part of #50.