BIG COMMIT. Introduce a Chunk class and use throughout.
The Chunk class is a wrapper/controller around the underlying data object in the Library._data['content'].

I likely broke something just given how big the change is, espeically for things that expect a raw dict as the chunk.

I did a smoke test before and after the change to verify that the normal host/client pipeline returns the same chunks before and after, but I likely broke something for example in the import
pipeline.

Part of #58.