Fix a bug where Library.slice() if sorted in order of similarity() would return a Library with a reverse sorting.
Slice inserts items back in one by one, so it was using _insert_chunk_into_ids.

_insert_chunk_into_ids was using bisect.bisect_left, but bisect and friends assume an ascending ordered list, and our sorts are reversed.

So, use a similarity that is reversed for bisect so that they are in the right order. (https://stackoverflow.com/a/69608002 was helpful).

This bug didn't actually influence the QUALITY of results, because they still had the right sliced chunks returned and the order in the context likely didn't matter, but the infos were in the reverse
order they should be.

Part of #14.