Change how Chunk.validate() works.
Previously it would be run at creation time. But you need the library to know which fields can and should be omitted.

Specifically in Library.extend(), we'd have copies of chunks who had no parent. They'd have their embedding omitted, which was OK by its soon-to-be-parent library, but they'd fail validation
erroneously when they were copied since they omitted a field that they weren't supposed to omit.

Part of #14.