Make it so we pick the best context from each server to pass to completion in sample.client
Before when we had multiple endpoints to query, we'd ask each one for only its fraction of its results.

This meant that if some endpoints had very good results and some had very poor results even the poor results would show up.

Now we merge all of the results into one library and keep the best results. As a result, the endpoints that have the most relevant content will dominate.

This significantly improves the quality of answers when querying multiple endpoints, but at the cost of a little less diversity of endpoints in which chunks are selected.

Part of #14.