Land basic upgrader machinery.
When a Library() is loaded, upgrade_library_data is called. It continually calls upgrader scripts that increment the version number up one by one until it goes as far as it knows how (implicitly
CURRENT_VERSION).

This allows clients to talk to older hosts without issue, because right when the library is loaded it is seemlessly upgraded into a current library version.

Part of #61.